rename internal to pkg
