+++ example/ion-sfu-gstreamer-receive/main.go
@@ -9,7 +9,7 @@ import (
 
 	log "github.com/pion/ion-log"
 	sdk "github.com/pion/ion-sdk-go"
-	gst "github.com/pion/ion-sdk-go/internal/gstreamer-sink"
+	gst "github.com/pion/ion-sdk-go/pkg/gstreamer-sink"
 	"github.com/pion/rtcp"
 	"github.com/pion/webrtc/v3"
 )
